third_party/WebKit/LayoutTests/fast/scrolling/scrolling-apis-nan-scroll-position-expected.txt

Verifies that nan scroll position works correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS window.scrollY is 1
PASS window.scrollX is 1
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 0
PASS window.scrollX is 0
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
PASS window.scrollX is 100
PASS scroller.scrollTop is 1
PASS scroller.scrollTop is 0
PASS scroller.scrollTop is 1
PASS scroller.scrollTop is 0
PASS scroller.scrollTop is 1
PASS scroller.scrollTop is 0
PASS scroller.scrollLeft is 1
PASS scroller.scrollLeft is 0
PASS scroller.scrollLeft is 1
PASS scroller.scrollLeft is 0
PASS scroller.scrollLeft is 1
PASS scroller.scrollLeft is 0
